ODDSvsEVENS
Coding Practice using java & arraylists created a Junit test file as well the objective of the challenge:
Given an integer, return "odd" if the sum of all odd digits is greater than the sum of all even digits. Return "even" if the sum of even digits is greater than the sum of odd digits, and "equal" if both sums are the same.